✅ Get the Maximum Score | Hard | C++ | Runtime: 0 ms | Memory: 12.2 MB | 2026-08-20

+++ leetcode/hard/1537-get-the-maximum-score/solution.cpp
@@ -9,7 +9,7 @@ public:
         int j=0;
         int n=nums1.size();
         int m=nums2.size();
-        int res=0;
+        ll res=0;
 
         while(i<n && j<m){
             if(nums1[i]<nums2[j]){
